feat: Introduce declarations `Decl` as a separate AST Node.
We break the old design of having declarations as part of statement
nodes by introducing a specialized node type `Decl` for global and
function declaration. This allows us to express the notion of programs
as sequence of declarations in a type safe way. Allowing us to remove
extraneous runtime scope checking during symbol table construction.

This also has the impact of introducing a new method to the `Visitor`
trait `visit_decl` and refactor the parser tests to specify inputs
as declarations. Parser tests could benefit from being more modular.