feat(sema): Drop `TypeAnalyzer` in favor of a `resolve`function.
`TypeAnalyzer` introduces an extra layer of lifetime management due to
it implementing the `Visitor` trait which takes a mutable reference to
`self`. This causes the lifetime of `TypeAnalyzer` to be ambiguous.

An alternative is to make `Visitor` take non-mutable reference to
`self` and use interior mutability via `RefCell` to mutate the state
of each visitor, but this has extra overhead since we borrow checking
will be done at runtime potentially causing unexpected behavior.

To solve these issues we just introduce `resolve` as a function on
the `SemanticAnalyzer` which does type resolution and ensures type
soundness of expressions.

As always, simple wins.