feat(ir): Start implementing intermediate representation phase.
The IR generation phase needs to be refactored into a format more
suitable implementation, changing the visitor signature to capture
both instruction and destination of various assignment is a good
first step.

Another thing that we may need to do is use a symbol pool to avoid
all allocations and string copies.